Nullkiller: rework defence a bit

+++ AI/Nullkiller/AIUtility.h
@@ -31,7 +31,7 @@ const int ACTUAL_RESOURCE_COUNT = 7;
 const int ALLOWED_ROAMING_HEROES = 8;
 
 //implementation-dependent
-extern const double SAFE_ATTACK_CONSTANT;
+extern const float SAFE_ATTACK_CONSTANT;
 extern const int GOLD_RESERVE;
 
 //provisional class for AI to store a reference to an owned hero object
